+++ HISTORY.rst
@@ -23,3 +23,4 @@ History
 * `60`: add notebook coverage as a separate page
 * `34`: applies pep8 on the code being rendered on the documentation
 * `65`: add function to clean readme.rst before sending it to pypi
+* `67`: add toctree delayed which gets filled after the dynamic content is created, use it for blogs